pyproject.toml
[tool.poetry]
name = "rent-crm"
version = "0.1.0"
description = ""
package-mode = false
authors = ["daniil_yarshou daniil_yarshou@epam.com"]



[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.12.5"
pydantic-settings = "^2.13.1"
fastapi = "^0.135.1"
alembic = "^1.18.4"
asyncpg = "^0.31.0"
sqlalchemy = "^2.0.48"
ruff = "^0.15.6"
isort = "^8.0.1"
uvicorn = "^0.41.0"
gunicorn = "^25.1.0"
pyyaml = "^6.0.3"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 125
indent-width = 4
exclude = [
    ".idea",
    ".venv",
    ".git",
    "alembic"
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
]
ignore = [
    "C901", # too complex
    "D401", # imperative mood
    "I"
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401", "F403"]

[tool.isort]
profile = "black"
line_length = 120

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
